skip if case-transform is not cloned

+++ modules/home-manager/editor/nixvim/plugins/case-transform.nix
@@ -1,19 +1,25 @@
 { pkgs, ... }:
 
+let
+  caseTransformPath = /home/takahisa/ghq/personal/github.com/amano-takahisa/case-transform.nvim;
+  caseTransformExists = builtins.pathExists caseTransformPath;
+in
 {
   programs.nixvim = {
-    extraPlugins = [
-      (pkgs.vimUtils.buildVimPlugin {
-        pname = "case-transform.nvim";
-        version = "dev";
-        src = /home/takahisa/ghq/personal/github.com/amano-takahisa/case-transform.nvim;
-        postPatch = "rm -rf doc";
-        meta.homepage = "https://github.com/amano-takahisa/case-transform.nvim";
-      })
-    ];
+    extraPlugins =
+      if caseTransformExists then [
+        (pkgs.vimUtils.buildVimPlugin {
+          pname = "case-transform.nvim";
+          version = "dev";
+          src = caseTransformPath;
+          postPatch = "rm -rf doc";
+          meta.homepage = "https://github.com/amano-takahisa/case-transform.nvim";
+        })
+      ] else [ ];
 
-    extraConfigLua = ''
-      require('case-transform').setup({ default_keymaps = true })
-    '';
+    extraConfigLua =
+      if caseTransformExists then ''
+        require('case-transform').setup({ default_keymaps = true })
+      '' else "";
   };
 }
